leadqual/requirements.txt
# LeadQual AI - Dependencies

# Core
python-dotenv>=1.0.0
openai>=1.0.0

# Database
psycopg2-binary>=2.9.0

# API Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# HTTP Requests (for Zoho/Gmail APIs)
httpx>=0.25.0
aiohttp>=3.9.0

# Authentication
pyjwt>=2.8.0
clerk-backend-api>=4.2.0

# Utilities
python-dateutil>=2.8.0

